backend/requirements.txt
aiosqlite==0.21.0
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncer==0.0.8
asyncpg==0.30.0
autoflake==2.3.1
bcrypt==4.3.0
billiard==4.2.1
black==25.1.0
bleach==6.2.0
cachetools==5.5.2
celery==5.3.6
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.8.0
cryptography==44.0.2
cssselect==1.3.0
cssutils==2.11.1
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
emails==0.6
exceptiongroup==1.2.2
factory_boy==3.3.3
Faker==37.1.0
fastapi==0.115.12
fastapi-async-sqlalchemy==0.6.1
fastapi-cache2==0.2.2
fastapi-cli==0.0.7
fastapi-csrf-protect==1.0.3
fastapi-limiter==0.1.6
fastapi-pagination==0.13.1
filelock==3.18.0
flake8==7.2.0
flake8-plugin-utils==1.3.3
flake8-pytest-style==2.1.0
flower==2.0.1
greenlet==3.2.1
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humanize==4.12.2
identify==2.6.10
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.3
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.7.0
mypy==1.11.2
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.2.5
orjson==3.10.16
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pendulum==3.1.0
pika==1.3.2
platformdirs==4.3.7
pluggy==1.5.0
pre-commit==3.8.0
premailer==3.10.0
prometheus_client==0.21.1
psycopg2-binary==2.9.9
prompt_toolkit==3.0.51
pyasn1==0.4.8
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.3
pydantic-extra-types==2.10.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.3
rsa==4.9.1
sentry-sdk==2.27.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlakeyset==2.0.1739066250
SQLAlchemy==2.0.40
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.24
starlette==0.46.2
tenacity==9.1.2
tomli==2.2.1
tornado==6.4.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
slowapi==0.1.9
uvicorn==0.34.2
vine==5.1.0
virtualenv==20.30.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
zxcvbn==4.5.0
